fix: Support fastify method chaining (#2)
Fastify supports method chaining when registering plugins and adding
decorators. The wrapped methods weren't returning the fastify instance
so any plugins that used chaining would throw errors during the start
of the application. This PR returns the fastify instance from the
`wrapDecorateMethod` and `wrapRegisterMethod` functions to now support
this feature.